+++ jwp-basic-gradle/src/main/java/next/controller/qna/ShowFormController.java
@@ -1,4 +1,4 @@
-package next.controller;
+package next.controller.qna;
 
 import core.mvcframework.Controller;
 import java.util.List;
