+++ jwp-basic-gradle/src/main/java/next/controller/user/CreateUserController.java
@@ -1,32 +1,28 @@
-package next.controller;
+package next.controller.user;
 
 import core.mvcframework.Controller;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
 import next.model.User;
 import next.service.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class UpdateUserController implements Controller {
-    private static final Logger log = LoggerFactory.getLogger(UpdateUserController.class);
+public class CreateUserController implements Controller {
+    private static final Logger log = LoggerFactory.getLogger(CreateUserController.class);
 
     private final UserService userService = new UserService();
 
     @Override
     public String execute(final HttpServletRequest request, final HttpServletResponse response) {
-        User updatedUser = new User(
+        User user = new User(
                 request.getParameter("userId"),
                 request.getParameter("password"),
                 request.getParameter("name"),
                 request.getParameter("email")
         );
-        userService.updateUserInformation(updatedUser);
-        log.info("updatedUser={}", updatedUser);
-        HttpSession session = request.getSession();
-        session.removeAttribute("user");
-        session.setAttribute("user", updatedUser);
-        return "redirect:/users/list";
+        userService.signUp(user);
+        log.debug("createUser={}", user);
+        return "redirect:/";
     }
 }
