+++ jwp-basic-gradle/src/main/java/next/controller/user/LogoutUserController.java
@@ -1,13 +1,14 @@
-package next.controller;
+package next.controller.user;
 
 import core.mvcframework.Controller;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class QuestionFormController implements Controller {
+public class LogoutUserController implements Controller {
 
     @Override
     public String execute(final HttpServletRequest request, final HttpServletResponse response) {
-        return "qna/form";
+        request.getSession().invalidate();
+        return "redirect:/";
     }
 }
